qr_generator/requirements.txt
pandas==2.2.3
Pillow==11.2.1
qrcode==8.2
reportlab==4.4.0
SQLAlchemy==2.0.41
streamlit==1.45.0
psycopg2-binary